Refactor API's and the impl.
